Move factory into separate file. #1486
---

Written with Claude Code. Here follows the plan.

---

New file: src/Commands/CreateEditorCallback.ts
- EditorInsertion interface (exported)
- LineTransformer type (exported)
- getNewCursorPosition function (exported)
- createEditorCallback factory (exported)

Updated file: src/Commands/ToggleDone.ts
- Import createEditorCallback, EditorInsertion from ./CreateEditorCallback
- Re-export getNewCursorPosition from ./CreateEditorCallback (to avoid breaking test imports)
- Keep toggleLine function
- Keep toggleDone = createEditorCallback(toggleLine)

Test file: tests/Commands/ToggleDone.test.ts
- No changes needed (re-export preserves the import path)
